Update with review comments
