+++ ledger/participant-integration-api/src/main/resources/metering-keys/community.json
@@ -1,0 +1,4 @@
+{
+  "algorithm": "HmacSHA256",
+  "encoded": "iENTFX4g-fAvOBTXnGjIVfesNzmWFKpo_35zpUnXEsg="
+}
