--- ledger/participant-integration-api/src/test/resources/keys/README.md
@@ -1,4 +1,0 @@
-
-# Metering Report Tamperproofing Keys
-
-This directory uses keys used to test tamperproofing
